refactor: Вынес протоколы репозиториев в дирректорию store
