+++ app/auth/store/repo/auth.py
@@ -1,6 +1,6 @@
-from auth.action.store.auth import IAuthUserRepository
-from auth.action.store.auth import ILoginUserRepository
 from auth.dto.register import StoreRegisterUserDTO
+from auth.store.proto.auth import IAuthUserRepository
+from auth.store.proto.auth import ILoginUserRepository
 from sqlalchemy.ext.asyncio.session import AsyncSession
 from sqlalchemy.sql import select
 from user.model.entity.user import User
