+++ app/infra/ioc/providers/auth/repo.py
@@ -1,6 +1,6 @@
-from auth.action.store.auth import IAuthUserRepository
-from auth.action.store.auth import ILoginUserRepository
-from auth.model.repo.auth import AuthUserRepository
+from auth.store.proto.auth import IAuthUserRepository
+from auth.store.proto.auth import ILoginUserRepository
+from auth.store.repo.auth import AuthUserRepository
 from dishka import provide
 from dishka import Provider
 from dishka import Scope
